fixes, remove !-sign

+++ src/app/features/home/login/login.component.ts
@@ -23,13 +23,15 @@ export class LoginComponent {
   }
 
   login() {
-    this.cookieService.deleteAll();
-    this.authenticationService.login(this.loginForm.value.email!, this.loginForm.value.password!).then(_ => {
-      this.router.navigateByUrl('/home');
-      this.dialog.closeAll();
-      console.log("van")
-    }).catch(value => {
-      this.errorMessage = value;
-    })
+    if (this.loginForm.value.email && this.loginForm.value.password) {
+      this.cookieService.deleteAll();
+      this.authenticationService.login(this.loginForm.value.email, this.loginForm.value.password).then(_ => {
+        this.router.navigateByUrl('/home');
+        this.dialog.closeAll();
+        console.log("van")
+      }).catch(value => {
+        this.errorMessage = value;
+      })
+    }
   }
 }
